refactor component file names
